Call reporter summary handler once from runner
Under certain conditions, the reporter's summary handler can be called
by the runner more than once. The default reporter handled such cases
correctly, but the runner can be configured with a user-provided
reporter type that may not align with that assumption. Moving the
responsibility to only call the reporter's summary handler once to the
runner aligns with user expectations and prevents bugs.

Issue: #515