Module build fixes (#568)
* Fixes some upstream build issues.

A using-declaration may not use an enumerator name.

* Removes an unneeded semicolon.

* Use named declarations of the std namespace.

The std module only provides named declarations of the C headers in the
std namespace. To use names in the global namespace the module
std.compat should be used.

* Use inline instead of static constexpr.

Modules may only export named declarations with external linkage.

* Slight modification in the usage of export.

The existing method doesn't work with Clang 17.