Implements streaming for invocable types. (#575)
Sometimes it's needed to write a custom message that can only be written
when the test fails. It's not possible to directly steam these messages,
instead lazy avaluation is needed. A typical example is to stream a
std::expected's error when it has no value. This is implemented by using
a callable that is only called when ut's expect fails.